+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2021-02-20
+
 ### Changed
 
+- **Breaking** - Change handles to localized strings
 - Pluralize image alt text collection
 
 ## [0.5.2] - 2021-01-31
@@ -46,8 +49,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Breaking** - Local entity interfaces, models and mappers, no longer needed
 - **Breaking** - ProductService and HealthService
-
-
 
 ## [0.4.2] - 2020-12-28
 
@@ -122,7 +123,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Health service using CQRS for status checks
 - In-memory backend providing default query handlers
 
-[unreleased]: https://github.com/SorenA/lightops-commerce-services-product/compare/0.5.2...develop
+[unreleased]: https://github.com/SorenA/lightops-commerce-services-product/compare/0.6.0...develop
+[0.6.0]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.6.0
 [0.5.2]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.5.2
 [0.5.1]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.5.1
 [0.5.0]: https://github.com/SorenA/lightops-commerce-services-product/tree/0.5.0